requirements.txt
streamlit
watchdog
openai
langchain
langchain_openai
langchain_community 
unstructured 
unstructured[pdf]
faiss-cpu
faiss-gpu
reportlab
huggingface_hub
pypdf2
python-docx